Map clients to stable ids

+++ src/server/server-options.zig
@@ -1,4 +1,4 @@
 pub const ServerOptions = struct {
     contracts: []const type,
-    max_clients: usize = 1024,
+    max_clients: u32 = 1024,
 };
